Importing code into a repository
I did not intend to use version control with this, as I did not think it
wouldd go anywhere. However, it would be easier on me and would poe no
downsides to put it on Github and Gitlab.

Currently, MPU configuration can be done with the static MPU class, and
the base of an ELF class has been started. A linked list has also been
implemented, but not yet used.